Move filter parsing into a singleton
